Fixed blocks in inventory turning into 2D and blocks with custom ItemBlock classes disappearing from world when upgrading a world from 1.6.
The issue with blocks turning into 2D happened because the block id was only getting re-mapped in the blockRegistry and not in the itemRegistry too which made Block.getBlockFromItem return air because Item.getIdFromItem was returning null, so it rendered in 2D.
And the issue with blocks with custom ItemBlock classes disappearing happened because FML was re-mapping it in the itemRegistry and not in the blockRegistry also.

I hope this makes sense and is correct.